+++ packages/pocl-job/src/__tests__/pocl-processor.spec.js
@@ -9,6 +9,10 @@ global.simulateLockError = false
 global.lockReleased = false
 jest.mock('@defra-fish/connectors-lib', () => ({
   ...jest.requireActual('@defra-fish/connectors-lib'),
+  airbrake: {
+    initialise: jest.fn(),
+    flush: jest.fn()
+  },
   DistributedLock: jest.fn().mockReturnValue({
     obtainAndExecute: jest.fn(async ({ onLockObtained, onLockError }) => {
       if (global.simulateLockError) {
@@ -66,11 +70,16 @@ describe('pocl-processor', () => {
     expect(processExitSpy).toHaveBeenCalledWith(0)
   })
 
-  it.each(['SIGINT', 'SIGTERM'])('implements a shutdown handler to respond to the %s signal', async signal => {
-    const processStopSpy = jest.spyOn(process, 'exit').mockImplementation(() => {})
-    await process.emit(signal)
-    expect(processStopSpy).toHaveBeenCalledWith(0)
-    expect(global.lockReleased).toEqual(true)
-    jest.restoreAllMocks()
+  describe.each(['SIGINT', 'SIGTERM'])('implements a shutdown handler to respond to the %s signal', signal => {
+    it('exits the process with code 0', done => {
+      const processStopSpy = jest.spyOn(process, 'exit').mockImplementation(jest.fn())
+      process.emit(signal)
+      setImmediate(() => {
+        expect(processStopSpy).toHaveBeenCalledWith(0)
+        expect(global.lockReleased).toEqual(true)
+        jest.restoreAllMocks()
+        done()
+      })
+    })
   })
 })